split nav bar
