+++ .eslintrc.js
@@ -18,6 +18,9 @@ module.exports = {
   plugins: ['react'],
   rules: {
     'react/jsx-filename-extension': [1, { extensions: ['.js', '.jsx'] }],
+    'no-unused-vars': 2,
+    'jsx-a11y/click-events-have-key-events': 0,
+    'jsx-a11y/no-static-element-interactions': 0,
     'no-use-before-define': 0,
     'arrow-parens': 0,
     'comma-dangle': 0
