+++ components/HorizontalSection/styles.scss
@@ -5,7 +5,6 @@ $color: red;
   overflow-x: scroll;
   width: auto;
   white-space: nowrap;
-  //   display: flex;
   &__item {
     display: inline-block;
     width: 100%;
