+++ package.json
@@ -15,7 +15,8 @@
     "next": "9.1.5",
     "node-sass": "^4.13.0",
     "react": "16.12.0",
-    "react-dom": "16.12.0"
+    "react-dom": "16.12.0",
+    "shortid": "^2.2.15"
   },
   "devDependencies": {
     "eslint": "^6.7.2",
